webserver/requirements.txt
flask-cors
numpy==1.16.5
Pillow==6.2.1
pymilvus==0.2.13
flask
flask_restful
gunicorn==20.0.0
futures==3.1.1
Keras==2.3.1
tensorflow==1.14.0
pymysql
